split a few components to their own files
